Fix printf format strings.
Note that I could have used PRId32 and friends but they are posix, and even
with the feature test macro I couldn't get them to work.  It doesn't matter
though, casts are perfectly adequate for this.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@10626 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime.17/src-cpp/x10aux/bootstrap.h
@@ -102,8 +102,8 @@ namespace x10aux {
             x10aux::ref<x10::lang::Throwable> &e_ =
                 static_cast<x10aux::ref<x10::lang::Throwable>&>(e);
 
-            fprintf(stderr, "Uncaught exception at place %ld: %s\n", x10aux::here(),
-                                                                     e_->toString()->c_str());
+            fprintf(stderr, "Uncaught exception at place %ld: %s\n", (long)x10aux::here(),
+                                                                    e_->toString()->c_str());
 
             e_->printStackTrace();
 
@@ -111,13 +111,11 @@ namespace x10aux {
 
         } catch(...) {
 
-            fprintf(stderr, "Caught unrecognised exception at place %ld\n", x10aux::here());
+            fprintf(stderr, "Caught unrecognised exception at place %ld\n", (long)x10aux::here());
             x10aux::exitCode = 1;
 
         }
 #endif
-
-        //fprintf(stderr, "Done with main in place %d", (int)x10aux::here());
 
         x10aux::free_args(args);
 
@@ -125,7 +123,7 @@ namespace x10aux {
 
         if (getenv("X10_RXTX")!=NULL)
             fprintf(stderr, "Place: %ld   rx: %lld   tx: %lld\n",
-                x10aux::here(),
+                (long)x10aux::here(),
                 (long long)x10aux::deserialized_bytes,
                 (long long)x10aux::serialized_bytes);
